re-organize imports so that rich is lazy-imported.
